simple game loop
A simple game loop is created to say how many frames we want per second and later how many updates per second we want. If a set duration has passed, it is time for a new frame and we need another refresh of the screen. In this case a repaint.
//repaint();